Allow custom creation and connection functions.
When using VSockets the code used to create a socket descriptor and connect it
to the target system are slightly different. This change adds a new connect()
method to Connection that allows callers to provide their own implementations
of these methods.

The existing inet connection code has been reimplemented to use this new
approach. The existing code used to allocate and connect inet connections has
been refactored into a pair of create and connect lambdas that are passed to
the new stuff.